Refactored person reid
